remove userId from PathVariable; commented index method in UserController; remove implement from UniversalController for UserController

+++ src/main/java/ru/coldwinternight/todo/controller/UserController.java
@@ -6,6 +6,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.web.bind.annotation.*;
+import ru.coldwinternight.todo.configuration.UserInfo;
 import ru.coldwinternight.todo.exception.IncorrectPasswordException;
 import ru.coldwinternight.todo.exception.UserAlreadyExistException;
 import ru.coldwinternight.todo.entity.UserEntity;
@@ -14,42 +15,45 @@ import ru.coldwinternight.todo.model.User;
 import ru.coldwinternight.todo.service.UserService;
 
 import javax.validation.Valid;
-import java.util.List;
 import java.util.Map;
 
 @Slf4j
 @RestController
 @RequiredArgsConstructor
 @RequestMapping("/users")
-public class UserController implements UniversalController {
+public class UserController{
     private final UserService userService;
+    private final UserInfo userInfo;
+
+//    @GetMapping
+//    public ResponseEntity<?> index() {
+//        try {
+//            log.info("Get all users");
+//            List<User> users = userService.readAll();
+//            return new ResponseEntity<>(users, HttpStatus.OK);
+//        } catch (UserNotFoundException | UsernameNotFoundException e) {
+//            log.error("Error, user not found: {}", e.getMessage());
+//            return new ResponseEntity<>(e.getMessage(), HttpStatus.NOT_FOUND);
+//        } catch (Exception e) {
+//            log.error("Error while getting all users : {}", e.getMessage());
+//            return new ResponseEntity<>(e.getMessage(), HttpStatus.BAD_REQUEST);
+//        }
+//    }
 
     @GetMapping
-    public ResponseEntity<?> index() {
+    public ResponseEntity<?> showById() {
         try {
-            log.info("Get all users");
-            List<User> users = userService.readAll();
-            return new ResponseEntity<>(users, HttpStatus.OK);
-        } catch (UserNotFoundException | UsernameNotFoundException e) {
-            log.error("Error, user not found: {}", e.getMessage());
-            return new ResponseEntity<>(e.getMessage(), HttpStatus.NOT_FOUND);
-        } catch (Exception e) {
-            log.error("Error while getting all users : {}", e.getMessage());
-            return new ResponseEntity<>(e.getMessage(), HttpStatus.BAD_REQUEST);
-        }
-    }
-
-    @GetMapping(value = "/{id}")
-    public ResponseEntity<?> showById(@PathVariable(name = "id") int id) {
-        try {
+            Integer id = userInfo.getUserId();
+            if (id == null)
+                throw new UserNotFoundException();
             log.info("Get one user with id {}", id);
             User user = userService.read(id);
             return new ResponseEntity<>(user, HttpStatus.OK);
         } catch (UserNotFoundException | UsernameNotFoundException e) {
-            log.error("Error while get {} user, user not found: {}", id, e.getMessage());
+            log.error("Error while get user: {}", e.getMessage());
             return new ResponseEntity<>(e.getMessage(), HttpStatus.NOT_FOUND);
         } catch (Exception e) {
-            log.error("Error while get {} user: {}", id, e.getMessage());
+            log.error("Error while get user: {}", e.getMessage());
             return new ResponseEntity<>(e.getMessage(), HttpStatus.BAD_REQUEST);
         }
     }
@@ -70,36 +74,43 @@ public class UserController implements UniversalController {
         }
     }
 
-    @PutMapping(value = "/{id}")
-    public ResponseEntity<?> update(@PathVariable(name = "id") int id, @Valid @RequestBody UserEntity user) {
+    @PutMapping
+    public ResponseEntity<?> update(@Valid @RequestBody UserEntity user) {
         try {
+            Integer id = userInfo.getUserId();
+            if (id == null)
+                throw new UserNotFoundException();
             log.info("Updating user {} with id {}", user, id);
             String updateMessage = "User successfully updated";
             userService.update(user, id);
             return new ResponseEntity<>(updateMessage, HttpStatus.OK);
         } catch (UserNotFoundException | UsernameNotFoundException e) {
-            log.error("Error while update user {}. User not found: {}", id, e.getMessage());
+            log.error("Error while update user. User not found: {}", e.getMessage());
             return new ResponseEntity<>(e.getMessage(), HttpStatus.NOT_FOUND);
         } catch (Exception e) {
-            log.error("Error while update user {}: {}", id, e.getMessage());
+            log.error("Error while update user: {}", e.getMessage());
             return new ResponseEntity<>(e.getMessage(), HttpStatus.NOT_MODIFIED);
         }
     }
 
-    @PatchMapping(value = "/{id}/password")
-    public ResponseEntity<?> updatePassword(@PathVariable(name = "id") int id,
-                                            @RequestBody Map<String, String> passwords) {
+    @PatchMapping("/password")
+    public ResponseEntity<?> updatePassword(@RequestBody Map<String, String> passwords) {
         String oldPassword, newPassword;
         String updatePasswordMessage = "Password was successfully updated.";
         try {
+            Integer id = userInfo.getUserId();
             log.info("Trying to change user's password. UserId: {}", id);
             oldPassword = passwords.get("oldpassword");
             newPassword = passwords.get("newpassword");
+
+            if (id == null)
+                throw new UserNotFoundException();
 
             if (oldPassword.equals(newPassword))
                 return new ResponseEntity<>("These passwords are equal.", HttpStatus.NOT_MODIFIED);
 
             userService.updatePassword(oldPassword, newPassword, id);
+            log.info("Password has changed for user {}", id);
             return new ResponseEntity<>(updatePasswordMessage, HttpStatus.OK);
         } catch (NullPointerException e) {
             log.error("Error while changing password, bad credentials: {}", e.getMessage());
@@ -117,9 +128,12 @@ public class UserController implements UniversalController {
         }
     }
 
-    @DeleteMapping("/{id}")
-    public ResponseEntity<?> delete(@PathVariable(name = "id") int id) {
+    @DeleteMapping
+    public ResponseEntity<?> delete() {
         try {
+            Integer id = userInfo.getUserId();
+            if (id == null)
+                throw new UserNotFoundException();
             log.info("Trying to delete user {}", id);
             String deleteMessage = "User successfully deleted";
             userService.delete(id);
@@ -133,9 +147,12 @@ public class UserController implements UniversalController {
         }
     }
 
-    @PatchMapping("/{id}/reverseRandomize")
-    public ResponseEntity<?> reverseRandomizeTasks(@PathVariable(name = "id") int id) {
+    @PatchMapping("/reverseRandomize")
+    public ResponseEntity<?> reverseRandomizeTasks() {
         try {
+            Integer id = userInfo.getUserId();
+            if (id == null)
+                throw new UserNotFoundException();
             log.info("Reverting randomize_today_tasks status");
             boolean status = userService.reverseRandomizeTasks(id);
             String statusReverseMessage = String.format("Randomize user tasks enabled: %b", status);
